[Refactor] Pagination 컴포넌트 통합
